+++ Graphs/chaco_parser.h
@@ -45,9 +45,9 @@ class ChacoParser : ParserInterface {
   std::ifstream input_file_;
   std::ifstream weight_file_;
 
-  size_t num_nodes_;
-  size_t num_edges_;
-  size_t num_entries_per_weight_vector_;
+  unsigned int num_nodes_;
+  unsigned int num_edges_;
+  unsigned int num_entries_per_weight_vector_;
   ChacoWeightMode node_weight_mode_;
   ChacoWeightMode edge_weight_mode_;
